initial migration for ScannerImpl
